replaced cast to int with cast to uintptr_t (on 64bit int is 4bytes and pointer is 8bytes and g++4 complains about it
svn path=/trunk/flamerobin/; revision=550

+++ src/myTreeCtrl.cpp
@@ -65,7 +65,7 @@ void myTreeCtrl::OnBeginDrag(wxTreeEvent& event)
         if (!m)
             return;
         wxString test;
-        test.Printf(wxT("OBJECT:%d"), (int)m);
+        test.Printf(wxT("OBJECT:%d"), (uintptr_t)m);
         wxTextDataObject textData(test);
         wxDropSource source(textData, this);
         source.DoDragDrop(wxDrag_AllowMove);
